Update env for pixi run
If binchicken is installed in a non-default pixi environment, it could be run with something like `pixi run -e binchicken binchicken build ...`. For some reason, pixi passes the `-e binchicken` down to the `pixi run --frozen --manifest-path ... conda env vars ...`, which then fails due to the internal pixi.toml not containing a binchicken env.

This may need to be done with all invocations of pixi run without a set env?